+++ frontend/src/authentication/dto/AuthenticationDto.ts
@@ -1,0 +1,6 @@
+import {UserInfoDto} from "./userInfo/UserInfoDto.ts";
+
+export interface AuthenticationDto {
+  readonly userInfo?: UserInfoDto,
+  readonly accessToken?: string
+}
