+++ frontend/src/authentication/dto/LoginRequestDto.ts
@@ -1,0 +1,4 @@
+export interface LoginRequestDto {
+  readonly email: string;
+  readonly password: string;
+}
