+++ frontend/src/authentication/dto/RegisterRequestDto.ts
@@ -1,0 +1,5 @@
+export interface RegisterRequestDto {
+  readonly username: string;
+  readonly email: string;
+  readonly password: string;
+}
