+++ frontend/src/authentication/dto/applicationUser/UserResponsePublicDto.ts
@@ -1,0 +1,4 @@
+export interface UserResponsePublicDto {
+    readonly userId: number,
+    readonly username: string
+}
