+++ frontend/src/authentication/dto/userAccount/AccountType.ts
@@ -1,0 +1,3 @@
+export enum AccountType {
+    LOCAL, GOOGLE, GITHUB, FACEBOOK
+}
