+++ frontend/src/authentication/dto/userAccount/UserAccountResponseDto.ts
@@ -1,0 +1,7 @@
+import {AccountType} from "./AccountType.ts";
+
+export interface UserAccountResponseDto {
+    readonly id: number,
+    readonly email: string,
+    readonly accountType: AccountType
+}
