+++ frontend/src/authentication/dto/userInfo/GlobalRole.ts
@@ -1,0 +1,3 @@
+export enum GlobalRole {
+  USER, ADMIN
+}
