+++ frontend/src/authentication/dto/userInfo/UserInfoDto.ts
@@ -1,0 +1,7 @@
+import {GlobalRole} from "./GlobalRole.ts";
+
+export interface UserInfoDto {
+  readonly username: string;
+  readonly email: string;
+  readonly roles: Set<GlobalRole>;
+}
