+++ frontend/src/common/context/NotificationProvider.tsx
@@ -1,23 +1,14 @@
 import React, {createContext, ReactNode, useContext} from "react";
-import Snackbar, {SnackbarOrigin} from "@mui/material/Snackbar";
-import {
-  Alert, AlertColor
-} from "@mui/material";
-
-export interface NotificationState extends SnackbarOrigin {
-  message: string;
-  open?: boolean;
-  type: AlertColor;
-  vertical: "top" | "bottom";
-  horizontal: "left" | "center" | "right";
-}
+import Snackbar from "@mui/material/Snackbar";
+import {Alert} from "@mui/material";
+import {NotificationStateDto} from "../dto/NotificationStateDto.ts";
 
 interface NotificationProviderProps {
   children: ReactNode;
 }
 
 interface NotificationContextType {
-  openNotification: (newState: NotificationState) => void;
+  openNotification: (newState: NotificationStateDto) => void;
 }
 
 const NotificationContext: React.Context<NotificationContextType> = createContext<NotificationContextType>({
@@ -28,7 +19,7 @@ const NotificationContext: React.Context<NotificationContextType> = createContex
 export function NotificationProvider({children}: NotificationProviderProps) {
   const autohideDuration = 6000;
 
-  const [notificationState, setNotificationState] = React.useState<NotificationState>({
+  const [notificationState, setNotificationState] = React.useState<NotificationStateDto>({
     open: false,
     type: "info",
     message: "",
@@ -38,9 +29,9 @@ export function NotificationProvider({children}: NotificationProviderProps) {
 
   /**
    *
-   * @param newState {NotificationState}
+   * @param newState {NotificationStateDto}
    */
-  const openNotification = (newState: NotificationState) => {
+  const openNotification = (newState: NotificationStateDto) => {
     setNotificationState({...newState, open: true});
   };
 
@@ -55,7 +46,10 @@ export function NotificationProvider({children}: NotificationProviderProps) {
     <NotificationContext.Provider value={{openNotification}}>
       <Snackbar open={notificationState.open}
                 autoHideDuration={autohideDuration}
-                anchorOrigin={{vertical: notificationState.vertical, horizontal: notificationState.horizontal}}
+                anchorOrigin={{
+                  vertical: notificationState.vertical,
+                  horizontal: notificationState.horizontal
+                }}
                 onClose={handleClose}>
         <Alert onClose={handleClose}
                severity={notificationState.type}>
