+++ frontend/src/index.tsx
@@ -4,7 +4,7 @@ import {AuthenticationProvider} from "./authentication/context/AuthenticationPro
 import {DevSupport} from "@react-buddy/ide-toolbox";
 import {ComponentPreviews, useInitial} from "./react_buddy_dev";
 import {ThemePaletteModeProvider} from "./theme/context/ThemePaletteModeProvider.tsx";
-import appRouter from "./config/routing/appRouter.tsx";
+import appRouter from "./routing/appRouter.tsx";
 import {RouterProvider} from "react-router-dom";
 import {NotificationProvider} from "./common/context/NotificationProvider.tsx";
