+++ frontend/src/routing/IMenuRoutes.ts
