+++ frontend/src/theme/context/IThemePaletteModeContext.ts
@@ -1,0 +1,6 @@
+import {PaletteMode} from "@mui/material";
+
+export interface IThemePaletteModeContext {
+  paletteMode: PaletteMode;
+  togglePaletteMode: () => void;
+}
